ENH: more refactoring, added assert_series_almost_equal
