+++ CHANGELOG.md
@@ -28,7 +28,8 @@ constructor. ([#6240](https://github.com/biocore/scikit-bio/issues/624))
 * Added phylogenetic diversity metrics, including weighted UniFrac, unweighted UniFrac, and Faith's Phylogenetic Diversity. These are accessible as ``skbio.diversity.beta.unweighted_unifrac``, ``skbio.diversity.beta.weighted_unifrac``, and ````skbio.diversity.alpha.faith_pd``, respectively.
 * Addition of function ``skbio.diversity.alpha_diversity``, which supports applying an alpha diversity metric to multiple samples in one call.
 * Addition of functions ``skbio.diversity.get_alpha_diversity_metrics`` and ``skbio.diversity.get_beta_diversity_metrics``, which allow for discovery of the alpha and beta diversity metrics implemented in scikit-bio.
-* ``TreeNode.to_array`` now supports replace ``nan`` branch lengths in the resulting branch length vector with the value provided as ``nan_length_value``. 
+* ``TreeNode.to_array`` now supports replace ``nan`` branch lengths in the resulting branch length vector with the value provided as ``nan_length_value``.
+* Added ``skbio.util.assert_series_almost_equal`` function for comparing ``pd.DataFrame`` objects in unit tests.
 
 ### Backward-incompatible changes [stable]
 * `Sequence.kmer_frequencies` now returns a `dict`. Previous behavior was to return a `collections.Counter` if `relative=False` was passed, and a `collections.defaultdict` if `relative=True` was passed. In the case of a missing key, the `Counter` would return 0 and the `defaultdict` would return 0.0. Because the return type is now always a `dict`, attempting to access a missing key will raise a `KeyError`. This change *may* break backwards-compatibility depending on how the `Counter`/`defaultdict` is being used. We hope that in most cases this change will not break backwards-compatibility because both `Counter` and `defaultdict` are `dict` subclasses.
