+++ skbio/diversity/_util.py
