+++ skbio/diversity/alpha/_base.py
@@ -14,7 +14,7 @@ from scipy.optimize import fmin_powell, minimize_scalar
 
 from skbio.stats import subsample_counts
 from skbio.util._decorator import experimental
-from skbio.diversity._validate import _validate_counts_vector
+from skbio.diversity._util import _validate_counts_vector
 
 
 @experimental(as_of="0.4.0")
